+++ src/app/components/util/pagination.jsx
@@ -14,11 +14,11 @@ function ChplPagination(props) {
     setRowsPerPage,
   } = props;
 
-  const handleChangePage = (event, newPage) => {
+  const handlePageChange = (event, newPage) => {
     setPage(newPage);
   };
 
-  const handleChangeRowsPerPage = (event) => {
+  const handleRowsPerPageChange = (event) => {
     setRowsPerPage(parseInt(event.target.value, 10));
     setPage(0);
   };
@@ -26,8 +26,8 @@ function ChplPagination(props) {
   return (
     <TablePagination
       component="div"
-      onChangePage={handleChangePage}
-      onChangeRowsPerPage={handleChangeRowsPerPage}
+      onPageChange={handlePageChange}
+      onRowsPerPageChange={handleRowsPerPageChange}
       count={count}
       page={page}
       rowsPerPage={rowsPerPage}